+++ src/main/java/uk/ac/ebi/ddi/pipeline/indexer/email/SubmissionCompleteSummary.java
@@ -1,12 +1,15 @@
 package uk.ac.ebi.ddi.pipeline.indexer.email;
 
 
+import lombok.Getter;
+
 /**
  * Summary of the submissions details, including PRIDE Inspector URL etc
  *
  * @author Yasset Perez-Riverol (ypriverol@gmail.com)
  * @date 18/08/2015
  */
+@Getter
 public class SubmissionCompleteSummary {
     private final String projectAccession;
     private final String doi;
@@ -30,33 +33,5 @@ public class SubmissionCompleteSummary {
         this.reviewerUserName = reviewerUserName;
         this.reviewerPassword = reviewerPassword;
         this.bioSamplesId = bioSamplesId;
-    }
-
-    public String getProjectAccession() {
-        return projectAccession;
-    }
-
-    public String getDoi() {
-        return doi;
-    }
-
-    public String getProjectName() {
-        return projectName;
-    }
-
-    public String getSubmitterName() {
-        return submitterName;
-    }
-
-    public String getReviewerUserName() {
-        return reviewerUserName;
-    }
-
-    public String getReviewerPassword() {
-        return reviewerPassword;
-    }
-
-    public String getBioSamplesId() {
-        return bioSamplesId;
     }
 }
