+++ src/main/java/uk/ac/ebi/ddi/pipeline/indexer/tasklet/generation/GenerateEBeyeExpressionAtlasXML.java
@@ -1,5 +1,7 @@
 package uk.ac.ebi.ddi.pipeline.indexer.tasklet.generation;
 
+import lombok.Getter;
+import lombok.Setter;
 import org.springframework.batch.core.StepContribution;
 import org.springframework.batch.core.scope.context.ChunkContext;
 import org.springframework.batch.repeat.RepeatStatus;
@@ -29,6 +31,8 @@ import java.util.List;
  *
  * Created by ypriverol (ypriverol@gmail.com) on 26/06/2016.
  */
+@Getter
+@Setter
 public class GenerateEBeyeExpressionAtlasXML extends AbstractTasklet {
 
     public static final Logger LOGGER = LoggerFactory.getLogger(GenerationArrayExpressXML.class);
@@ -56,33 +60,7 @@ public class GenerateEBeyeExpressionAtlasXML extends AbstractTasklet {
     @Override
     public void afterPropertiesSet() throws Exception {
         Assert.notNull(outputFile, "Output directory cannot be null.");
-
         Assert.notNull(experimentFileName, "experiment prefix can't be null.");
         Assert.notNull(geneFileName,   "protocolPrefix can't be null.");
-
-    }
-
-    public String getOutputFile() {
-        return outputFile;
-    }
-
-    public void setOutputFile(String outputFile) {
-        this.outputFile = outputFile;
-    }
-
-    public String getExperimentFileName() {
-        return experimentFileName;
-    }
-
-    public void setExperimentFileName(String experimentFileName) {
-        this.experimentFileName = experimentFileName;
-    }
-
-    public String getGeneFileName() {
-        return geneFileName;
-    }
-
-    public void setGeneFileName(String geneFileName) {
-        this.geneFileName = geneFileName;
     }
 }
