+++ src/main/java/uk/ac/ebi/ddi/pipeline/indexer/tasklet/io/CleanDirectoryTasklet.java
@@ -1,5 +1,7 @@
 package uk.ac.ebi.ddi.pipeline.indexer.tasklet.io;
 
+import lombok.Getter;
+import lombok.Setter;
 import org.springframework.batch.core.StepContribution;
 import org.springframework.batch.core.scope.context.ChunkContext;
 import org.springframework.batch.repeat.RepeatStatus;
@@ -12,6 +14,8 @@ import uk.ac.ebi.ddi.pipeline.indexer.tasklet.AbstractTasklet;
  * @author Yasset Perez-Riverol (ypriverol@gmail.com)
  * @date 05/05/2016
  */
+@Getter
+@Setter
 public class CleanDirectoryTasklet extends AbstractTasklet {
 
     Resource inputDirectory;
@@ -25,13 +29,5 @@ public class CleanDirectoryTasklet extends AbstractTasklet {
     @Override
     public void afterPropertiesSet() throws Exception {
         Assert.notNull(inputDirectory, "Target directory can not be null");
-    }
-
-    public Resource getInputDirectory() {
-        return inputDirectory;
-    }
-
-    public void setInputDirectory(Resource inputDirectory) {
-        this.inputDirectory = inputDirectory;
     }
 }
