--- src/main/java/uk/ac/ebi/ddi/pipeline/indexer/utils/Tuple.java
@@ -1,57 +1,0 @@
-package uk.ac.ebi.ddi.pipeline.indexer.utils;
-
-import java.io.Serializable;
-import java.util.Objects;
-
-/**
- * @author Yasset Perez-Riverol (ypriverol@gmail.com)
- * @date 18/08/2015
- */
-
-public class Tuple<K, V> implements Serializable {
-    private K key;
-    private V value;
-
-    public Tuple(K key, V value) {
-        this.key = key;
-        this.value = value;
-    }
-
-    public K getKey() {
-        return key;
-    }
-
-    public void setKey(K key) {
-        this.key = key;
-    }
-
-    public V getValue() {
-        return value;
-    }
-
-    public void setValue(V value) {
-        this.value = value;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (!(o instanceof Tuple)) {
-            return false;
-        }
-
-        Tuple tuple = (Tuple) o;
-
-        return Objects.equals(key, tuple.key) && Objects.equals(value, tuple.value);
-
-    }
-
-    @Override
-    public int hashCode() {
-        int result = key != null ? key.hashCode() : 0;
-        result = 31 * result + (value != null ? value.hashCode() : 0);
-        return result;
-    }
-}
